// src/lib/types/dashboard-statsTypes.ts
export interface TotalAssessmentsResponse {
  totalAssessments: number;
  status: string;
}

export interface OngoingAssessmentsResponse {
  totalAssessments: number;
  status: string;
}

export interface ScheduledAssessmentsResponse {
  totalAssessments: number;
  status: string;
}

export interface PastAssessmentsResponse {
  totalAssessments: number;
  status: string;
}

export interface DraftAssessmentsResponse {
  totalAssessments: number;
  status: string;
}